Restructuring files for Cleaner Architecture
UseCases folder has different use case folders under it. Each related use cases now exist under one folder. This structure is important for inputBoundary files.
