+++ server/src/Controllers/getIntents.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const router = express.Router()
 const auth = require('../Middleware/auth')
-const getIntentsInteractor = require('../Use Cases/getIntents')
+const getIntentsInteractor = require('../UseCases/IntentUseCases/getIntents')
 
 router.get('/getIntents', auth, async (req, res) => {
     const user = req.user
