+++ server/src/Controllers/getOneTranscriptIntents.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const router = express.Router()
 const auth = require('../Middleware/auth')
-const getOneTranscriptIntentsInteractor = require('../Use Cases/getOneTranscriptIntents')
+const getOneTranscriptIntentsInteractor = require('../UseCases/TranscriptUseCases/getOneTranscriptIntents')
 
 router.get('/getOneTranscriptIntents', auth, async (req, res) => {
     const user = req.user
