+++ server/src/Controllers/getTranscripts.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const router = express.Router()
 const auth = require('../Middleware/auth')
-const getTranscriptsInteractor = require('../Use Cases/getTranscripts')
+const getTranscriptsInteractor = require('../UseCases/TranscriptUseCases/getTranscripts')
 
 router.get('/getTranscripts', auth, async (req, res) => {
     const user = req.user
