+++ server/src/UseCases/IntentUseCases/getIntents.js
