+++ server/src/UseCases/utils/transcriptProcessor.js
