+++ server/src/express.js
@@ -1,9 +1,9 @@
 const express = require('express')
 const cors = require('cors')
 
-const createUserController = require('./Controllers/createUser')
+const createUserController = require('./Controllers/userRegisterController')
 const loginUserController = require('./Controllers/loginUser')
-const uploadTranscriptController = require('./Controllers/uploadTranscript')
+const uploadTranscriptController = require('./Controllers/transcriptUploadController')
 const getIntentsController = require('./Controllers/getIntents')
 const getOneTranscriptIntentsController = require('./Controllers/getOneTranscriptIntents')
 const getTranscriptsController = require('./Controllers/getTranscripts')
